Add new command "!" which repeats previous command. Mention "!" option in "help" output.
When Pico W is coming on-line, if the chat service is unable to
bind to a socket (maybe because the service was just running)
the chat serviec will wait loop, waiting ten seconds and then
trying to bind to the socket again. This saves us from restarting
the Pico (and robot) each time a socket bind fails.